// frontend/src/app/api/model/habitFrequency.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const HabitFrequency = {
    Daily: 'DAILY',
    Weekly: 'WEEKLY',
} as const;
export type HabitFrequency = typeof HabitFrequency[keyof typeof HabitFrequency];