// frontend/src/app/api/model/plan.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanItem } from './planItem';
import { PlanStatus } from './planStatus';


export interface Plan { 
    id: number;
    title: string;
    estimatedDurationMinutes: number;
    startDateTime: string;
    endDateTime: string;
    priorityOrder: number;
    status: PlanStatus;
    createdAt: string;
    startNotifiedAt?: string | null;
    items: Array<PlanItem>;
    itemsTotal: number;
    itemsDone: number;
    progressPercent: number;
    restSeconds?: number | null;
}
export namespace Plan {
}